requirements.txt
aiogram>=3.0.0
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
shapely>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.25.0
